Camera functionality on home screen
When a user presses the plus button on the home screen, a modul will popup on the bottom of the screen prompting them if they want to take a picture or use one from the gallery. Also implemented code to display the taken or chosen picture on the home screen but its not working 100% right  now.