Added imageCache.clear() to both get image functions
imageCache.clear() is called every time a new picture is taken or picked from the gallery. It clears the image cache in flutter so flutter doesnt try to load the image from a cache. Instead it will always load the one that was just taken or picked. I haven't figured out yet if it needs to be in the setState() or not.