Finishing Authentication
Created pages for sign in and register. User can toggle between these two pages. Users can register and account and sign in with the account. Upon signing in, the user is automatically redirected to the home page. The is validation on each text field when signing in or registering to make sure the user actually input something. Need to add password check field to make sure passwords match.