Database functionality for profile page and creation of createPost page
Profile page now gets data from the database. It will grab the users data from the UserData collection via a function in the database.dart file that was added. The profile page now displays the users information and the number of posts they have created. The createPost page was made. This widget is loaded as soon as the user takes the first picture or selects the first picture. Currently it displays the picture and a textbox for the title and gets the user's username from the database. There is an issue where the username is not grabbed when the app auto signs in.