Postmaster and Demopost classes
Created the Postmaster class which holds all the demopost objects for storage.
Created the demopost class that holds onto title and image links
Added the deodata.txt which stores the names and links of posts, although I will probably change this to something else because its giving me trouble.
Added assets to the yaml file