Multiple picture posts
Functionality for posts with multiple pictures works. There is now a button with a plus that allows the user to add another picture to the post. The newly taken picture is added to the screen and pushes the add button next to it. These additional pictures are now also added to the data base.